# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation ",
        "description": "API Documentation",
        "contact": {
            "email": "srundavith@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get List Artists Data",
                "description": "enter your Artists here",
                "operationId": "Artists",
                "responses": {
                    "default": {
                        "description": "return array model Artists"
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "artists",
                "description": "artists",
                "operationId": "artists",
                "requestBody": {
                    "description": "form artists",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "artist_name",
                                    "artist_profile"
                                ],
                                "properties": {
                                    "artist_name": {
                                        "type": "string"
                                    },
                                    "artist_profile": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/api/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create a new payment",
                "description": "Create a new payment and handle subscription using Stripe Payment Intents.",
                "operationId": "createPayment",
                "requestBody": {
                    "description": "Payment data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "card_number",
                                    "expiry",
                                    "cvv",
                                    "subscription_plan_id"
                                ],
                                "properties": {
                                    "card_number": {
                                        "description": "Credit card number",
                                        "type": "string"
                                    },
                                    "expiry": {
                                        "description": "Expiration date in MM/YY format",
                                        "type": "string"
                                    },
                                    "cvv": {
                                        "description": "CVV code",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "Name on the card",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "subscription_plan_id": {
                                        "description": "ID of the subscription plan",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model payments"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "payments/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete payments",
                "description": "-",
                "operationId": "payments/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}